refactor(map): type the draw.* events in one place
`draw.*` events come from @mapbox/mapbox-gl-draw, not MapLibre, so they
are absent from MapEventType and every on/off/fire for them is a type
error. DrawControl already confined that cast to a local helper; the
fire("draw.update") in BabsIconController was still erroring.

Moves the cast into views/map/drawEvents.ts and adds fireDrawEvent, so
both call sites share it and there is one place to delete if the upstream
typings ever gain these events.
